refactor(run): move run tooling to pnpm
