+++ run/package.json
@@ -30,28 +30,25 @@
   "typings": "lib/cjs/index.d.ts",
   "types": "lib/cjs/index.d.ts",
   "dependencies": {
-    "xstream": "10.x || 11.x",
-    "quicktask": "1.1.0"
+    "quicktask": "1.1.0",
+    "xstream": "10.x || 11.x"
   },
   "devDependencies": {
-    "@types/es6-promise": "0.0.x",
-    "@types/mocha": "2.2.x",
-    "@types/node": "7.0.x",
-    "@types/sinon": "1.16.x",
+    "@types/mocha": "5.2.x",
+    "@types/node": "10.5.x",
+    "@types/sinon": "5.0.x",
+    "mocha": "^5.2.0",
+    "sinon": "^6.1.4",
+    "ts-node": "^7.0.0",
+    "typescript": "~2.8.4",
     "xstream": "11.x"
   },
   "engines": {
     "node": ">=0.12.0"
   },
-  "browserify-shim": {
-    "xstream": "global:xstream"
-  },
   "scripts": {
-    "mocha": "../node_modules/.bin/mocha test/*.ts --require ts-node/register --exit",
-    "test": "npm run mocha",
+    "test": "mocha test/*.ts --require ts-node/register --exit",
     "test-ci": "npm run test",
-    "browserify": "../node_modules/.bin/browserify lib/cjs/index.js --global-transform=browserify-shim --standalone Cycle --exclude xstream -o dist/cycle-run.js",
-    "minify": "node ../.scripts/minify.js dist/cycle-run.js dist/cycle-run.min.js",
     "postlib": "cp lib/cjs/adapt.* lib/"
   },
   "publishConfig": {
